Refactor: #3 user skeleton에 맞춰서 기존 코드 이동
