+++ src/main/java/com/zerobase/munbanggu/config/auth/SessionUser.java
@@ -1,6 +1,6 @@
 package com.zerobase.munbanggu.config.auth;
 
-import com.zerobase.munbanggu.model.entity.User;
+import com.zerobase.munbanggu.user.model.entity.User;
 import java.io.Serializable;
 
 public class SessionUser implements Serializable {
@@ -12,7 +12,7 @@ public class SessionUser implements Serializable {
     public SessionUser(User user) {
         this.name = user.getName();
         this.email = user.getEmail();
-        this.profile = user.getProfile();
+        this.profile = user.getProfileImageUrl();
     }
 
 }
