--- src/main/java/com/zerobase/munbanggu/repository/UserRepository.java
@@ -1,10 +1,0 @@
-package com.zerobase.munbanggu.repository;
-
-import com.zerobase.munbanggu.model.entity.User;
-import java.util.Optional;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-
-    Optional<User> findByEmail(String email);
-}
