--- src/main/java/com/zerobase/munbanggu/type/AuthProvider.java
@@ -1,6 +1,0 @@
-package com.zerobase.munbanggu.type;
-
-public enum AuthProvider {
-
-    KAKAO
-}
