+++ src/main/java/com/zerobase/munbanggu/user/repository/UserRepository.java
@@ -1,10 +1,13 @@
 package com.zerobase.munbanggu.user.repository;
 
 import com.zerobase.munbanggu.user.model.entity.User;
+import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
 public interface UserRepository extends JpaRepository<User, Long> {
+
+    Optional<User> findByEmail(String email);
 
 }
