+++ src/main/resources/application.yml
@@ -1,7 +1,7 @@
 spring:
   datasource:
     driver-class-name: org.h2.Driver
-    url: jdbc:h2:mem:munbanggu;MODE=MySQL
+    url: jdbc:h2:mem:munbanggu;NON_KEYWORDS=USER;
     username: sa
     password:
   h2:
